Use forkjoin in lieu of merge for combining spotify search results.
LGTM EK
Closes #17

Previously we were using merge which I believe was incorrect. It
returned results from the service in piecemeal and was dependent on
scheduling so was never completing properly. Also, a class level
variable was being used for the search observable in the search method
which was likely making things worse. This is much cleaner.